Refactorizacion de las vistas en base a la refactorizacion de los hooks y el AuthContext
