+++ client/src/components/AdminRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '../context/AuthContext';
+import useAuth from '../hooks/useAuth';
 
 export function AdminRoute() {
   const { user } = useAuth();
